backend-api/requirements.txt
Flask==2.1.3
Jinja2==3.1.0
MarkupSafe==2.1.0
pillow>=9.5,<10
boto3>=1.28.57
botocore>=1.31.57
awscli>=1.29.57
Flask-CORS>=3.1.1
